fix: resolve Traditional Chinese navigation links
